requirements.txt
django
djangorestframework
psycopg[binary]
celery
redis
python-dotenv
django-cors-headers
dj-database-url
gunicorn
